Implemented jpa only

+++ BulkSMS/src/main/java/co/ke/bigfootke/app/pojos/DispatchedSms.java
@@ -6,7 +6,7 @@ public class DispatchedSms {
 
 	private String message;
 	private List<Long> groupIds;
-	private Double cost;
+	private int cost;
 	private Long senderId;
 	
 	public DispatchedSms() {
@@ -28,11 +28,11 @@ public class DispatchedSms {
 		this.groupIds = groupIds;
 	}
 
-	public Double getCost() {
+	public int getCost() {
 		return cost;
 	}
 
-	public void setCost(Double cost) {
+	public void setCost(int cost) {
 		this.cost = cost;
 	}
